move BottomSheet component to a common place
