+++ src/app/Components/ArtworkLists/views/SelectArtworkListsForArtworkView/SelectArtworkListsForArtworkView.tsx
@@ -1,6 +1,5 @@
 import { Flex } from "@artsy/palette-mobile"
 import { useArtworkListsContext } from "app/Components/ArtworkLists/ArtworkListsContext"
-import { AutomountedBottomSheetModal } from "app/Components/ArtworkLists/components/AutomountedBottomSheetModal"
 import { SelectArtworkListsForArtwork } from "app/Components/ArtworkLists/views/SelectArtworkListsForArtworkView/components/SelectArtworkListsForArtwork"
 import { SelectArtworkListsForArtworkHeader } from "app/Components/ArtworkLists/views/SelectArtworkListsForArtworkView/components/SelectArtworkListsForArtworkHeader"
 import {
@@ -8,6 +7,7 @@ import {
   StickyBottomContentPlaceholder,
 } from "app/Components/ArtworkLists/views/SelectArtworkListsForArtworkView/components/StickyBottomContent"
 import { ArtworkListsViewName } from "app/Components/ArtworkLists/views/constants"
+import { AutomountedBottomSheetModal } from "app/Components/BottomSheet/AutomountedBottomSheetModal"
 
 const SNAP_POINTS = ["50%", "95%"]
