+++ src/app/Scenes/ArtworkList/views/DeleteArtworkListView/DeleteArtworkListView.tsx
@@ -3,12 +3,12 @@ import { Button, Flex, Spacer, Text } from "@artsy/palette-mobile"
 import { useBottomSheetModal } from "@gorhom/bottom-sheet"
 import { captureMessage } from "@sentry/react-native"
 import { ArtworkListsBottomSheetSectionTitle } from "app/Components/ArtworkLists/components/ArtworkListsBottomSheetSectionTitle"
+import { useArtworkListsBottomOffset } from "app/Components/ArtworkLists/useArtworkListsBottomOffset"
+import { useArtworkListToast } from "app/Components/ArtworkLists/useArtworkListsToast"
 import {
   AutoHeightBottomSheet,
   AutoHeightBottomSheetProps,
-} from "app/Components/ArtworkLists/components/AutoHeightBottomSheet"
-import { useArtworkListsBottomOffset } from "app/Components/ArtworkLists/useArtworkListsBottomOffset"
-import { useArtworkListToast } from "app/Components/ArtworkLists/useArtworkListsToast"
+} from "app/Components/BottomSheet/AutoHeightBottomSheet"
 import { HeaderMenuArtworkListEntity } from "app/Scenes/ArtworkList/types"
 import { goBack } from "app/system/navigation/navigate"
 import { FC } from "react"
